chore: move get_clock to bench
