+++ governance/program/tests/process_cancel_proposal.rs
@@ -33,7 +33,7 @@ async fn test_cancel_proposal() {
         .await
         .unwrap();
 
-    let clock = governance_test.get_clock().await;
+    let clock = governance_test.bench.get_clock().await;
 
     // Act
     governance_test
