+++ governance/program/tests/process_flag_instruction_error.rs
@@ -62,7 +62,7 @@ async fn test_execute_flag_instruction_error() {
         .advance_clock_by_min_timespan(proposal_instruction_cookie.account.hold_up_time as u64)
         .await;
 
-    let clock = governance_test.get_clock().await;
+    let clock = governance_test.bench.get_clock().await;
 
     // Act
     governance_test
